requirements.txt
pyjwt==2.10.1
bcrypt==5.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
sqlalchemy==2.0.44
psycopg[binary]==3.2.3
python-dateutil==2.9.0
python-dotenv==1.0.1